Products should be updated to route variant
