+++ frontend/components.d.ts
@@ -36,17 +36,19 @@ declare module 'vue' {
     ConfirmationDialog: typeof import('./src/components/common/ConfirmationDialog.vue')['default']
     CreateBoxCard: typeof import('./src/components/widgets/Create/CreateBoxCard.vue')['default']
     CreateCard: typeof import('./src/components/widgets/Create/CreateCard.vue')['default']
-    CreateHouseholdCard: typeof import('./src/components/widgets/Households/CreateHouseholdCard.vue')['default']
+    CreateHouseholdCard: typeof import('./src/components/widgets/Households/Card/CreateHouseholdCard.vue')['default']
     CreateLocationCard: typeof import('./src/components/widgets/Create/CreateLocationCard.vue')['default']
     CreateProductCard: typeof import('./src/components/widgets/Create/CreateProductCard.vue')['default']
+    EditHousehold: typeof import('./src/components/widgets/Households/EditHousehold.vue')['default']
     ForgotPasswordCard: typeof import('./src/components/widgets/Login/ForgotPasswordCard.vue')['default']
     GeneralSettings: typeof import('./src/components/widgets/settings/GeneralSettings.vue')['default']
-    HouseholdCard: typeof import('./src/components/widgets/Households/HouseholdCard.vue')['default']
-    HouseholdInvite: typeof import('./src/components/widgets/Households/HouseholdInvite.vue')['default']
+    HouseholdCard: typeof import('./src/components/widgets/Households/Card/HouseholdCard.vue')['default']
+    HouseholdInvite: typeof import('./src/components/widgets/Households/Card/HouseholdInvite.vue')['default']
     HouseholdItem: typeof import('./src/components/household/HouseholdItem.vue')['default']
     HouseholdMember: typeof import('./src/components/widgets/Households/HouseholdMember.vue')['default']
-    HouseholdMemberCard: typeof import('./src/components/widgets/Households/HouseholdMemberCard.vue')['default']
-    HouseholdQuickSelectCard: typeof import('./src/components/widgets/Households/HouseholdQuickSelectCard.vue')['default']
+    HouseholdMemberCard: typeof import('./src/components/widgets/Households/Card/HouseholdMemberCard.vue')['default']
+    HouseholdQuickSelectCard: typeof import('./src/components/widgets/Households/Card/HouseholdQuickSelectCard.vue')['default']
+    Households: typeof import('./src/components/widgets/Households/Households.vue')['default']
     InviteModal: typeof import('./src/components/widgets/Households/InviteModal.vue')['default']
     IroColorPicker: typeof import('./src/components/ui/IroColorPicker.vue')['default']
     LabelSettings: typeof import('./src/components/widgets/QrCode/Print/LabelSettings.vue')['default']
@@ -62,13 +64,16 @@ declare module 'vue' {
     PrintQrCodeSettings: typeof import('./src/components/widgets/QrCode/Print/PrintQrCodeSettings.vue')['default']
     PrintQrCodesModal: typeof import('./src/components/widgets/QrCode/Print/PrintQrCodesModal.vue')['default']
     PrintQrModal: typeof import('./src/components/widgets/QrCode/Print/PrintQrModal.vue')['default']
-    ProductCard: typeof import('./src/components/widgets/products/ProductCard.vue')['default']
+    Product: typeof import('./src/components/widgets/products/Product.vue')['default']
+    ProductCard: typeof import('./src/components/widgets/products/Cards/ProductCard.vue')['default']
     ProductOverlay: typeof import('./src/components/widgets/products/ProductOverlay.vue')['default']
-    ProductStorageCard: typeof import('./src/components/widgets/products/ProductStorageCard.vue')['default']
+    Products: typeof import('./src/components/widgets/products/Products.vue')['default']
+    ProductStorage: typeof import('./src/components/widgets/products/ProductStorage.vue')['default']
+    ProductStorageCard: typeof import('./src/components/widgets/products/Cards/ProductStorageCard.vue')['default']
     ProfileCard: typeof import('./src/components/widgets/Account/ProfileCard.vue')['default']
     QrCodeFilter: typeof import('./src/components/common/QrCodeFilter.vue')['default']
     QrScanner: typeof import('./src/components/widgets/QrCode/Scan/QrScanner.vue')['default']
-    QuickActions: typeof import('./src/components/widgets/products/QuickActions.vue')['default']
+    QuickActions: typeof import('./src/components/widgets/products/Cards/QuickActions.vue')['default']
     RegisterCard: typeof import('./src/components/widgets/Register/RegisterCard.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
