+++ frontend/src/api/http/ProductEndpoint.ts
@@ -1,6 +1,7 @@
 import {Endpoint} from "./Endpoint.ts";
 import {ApiProduct, ProductStorageMapping} from "@/types";
 
+
 export class ProductEndpoint extends Endpoint{
 
     constructor(){
@@ -16,7 +17,6 @@ export class ProductEndpoint extends Endpoint{
             url += `/${id}`
         }
         const response = await this.internalAxios.get(url)
-
         if(response.status === 200){
             return response.data as Array<ApiProduct>
         }
