+++ frontend/src/components/widgets/Households/Card/HouseholdInvite.vue
