+++ frontend/src/components/widgets/Households/Card/HouseholdMemberCard.vue
