+++ frontend/src/components/widgets/Households/Card/HouseholdQuickSelectCard.vue
