+++ frontend/src/components/widgets/products/Cards/ProductCard.vue
