+++ frontend/src/components/widgets/products/Cards/QuickActions.vue
