+++ frontend/src/global.scss
@@ -63,5 +63,6 @@ input[type=number] {
     right: 0;
     .scroll{
         height: 100%;
+        scroll-behavior: smooth;
     }
 }
