+++ frontend/src/router/routes/household.ts
@@ -1,12 +1,25 @@
 import Household from "@/views/Household.vue";
 import {i18n} from "@/lang";
-import EditHousehold from "@/views/EditHousehold.vue";
+import EditHousehold from "@/components/widgets/Households/EditHousehold.vue";
+import Households from "@/components/widgets/Households/Households.vue";
 
-
-export const householdsOverview = {
+export default  {
     path: "/households",
     name: "households",
     component: Household,
+    children: [
+        {
+            path: '',
+            name: "all households",
+            component: Households,
+        },
+        {
+            path: "edit/:id",
+            name: "edit household",
+            component: EditHousehold,
+            props: true,
+        }
+    ],
     meta: {
         requiresAuth: true,
         fillHeight: true,
@@ -14,14 +27,3 @@ export const householdsOverview = {
     }
 }
 
-export const editHousehold = {
-    path: "/household/edit/:id",
-    name: "edit household",
-    component: EditHousehold,
-    props: true,
-    meta: {
-        requiresAuth: true,
-        fillHeight: true,
-        title: i18n.global.t('titles.edit_household')
-    }
-}
